Using blocking for now
After spending a full day fighting mocking and async, I decided to make
things simple and just start with the blocking interface of reqwest for
now.

Tokio is something I don't fully understand and it's bringing more
problems than advantages at the moment. Hopefully, I will be able to
convert the project to use async when things are clearer in the future.