Use the release version and free memory
For the example it's not really needed, but the library itself needs a
way to free a string.

It's a common pattern to let the creator of the memory to free it.
Freeing the memory from a different language may cause some unwanted
behaviors. The FFI gem actually has a good way to deal with that via
`FFI::AutoPointer` but it will not be implemented for this example right
now.